fixed setup bugs, gamefield ids are now stored as integers

+++ src/main/kotlin/net/spacetivity/blocko/field/GameFieldDAO.kt
@@ -6,8 +6,8 @@ import org.jetbrains.exposed.sql.Table
 object GameFieldDAO : Table("game_fields") {
     val arenaId = arenaId("arenaId", 10)
     val worldName = varchar("worldName", 30)
-    val x = double("x")
-    val z = double("z")
+    val x = integer("x")
+    val z = integer("z")
     val properties = text("properties")
     val isGarageField = bool("isGarage")
 }
